+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - add date/time support
 - add `&` (bitwise and) and `|` (bitwise or) operators
 - fix incompatibility with 'mathn' module
+- add `CONTAINS` string function
 
 ## [v2.0.10] 2016-12-30
 - fix string function initialization bug